+++ lib/src/lib.rs
@@ -62,6 +62,7 @@ pub mod simple_op_heads_store;
 pub mod simple_op_store;
 pub mod stacked_table;
 pub mod store;
+pub mod str_util;
 pub mod submodule_store;
 pub mod transaction;
 pub mod tree;
